refactor(offline-sync): move parseOfflineSyncExcludes next to the glob compiler
config.ts carries only the key wiring (+5 lines over baseline: import +
validated key + comment) — justification: new user-facing config key
(#1786); validation logic lives in offline-sync.ts beside
compileOfflineSyncExcludeGlobs. Ratchet baseline updated accordingly.
