+++ scripts/ratchet-baseline.json
@@ -6,9 +6,9 @@
     "watchlistLoc": {
       "packages/remnic-core/src/orchestrator.ts": 14426,
       "packages/remnic-core/src/cli.ts": 9395,
-      "packages/remnic-core/src/access-service.ts": 9017,
+      "packages/remnic-core/src/access-service.ts": 9039,
       "packages/remnic-core/src/storage.ts": 7747,
-      "packages/remnic-core/src/config.ts": 4689
+      "packages/remnic-core/src/config.ts": 4694
     },
     "oversizedFileCount": 10,
     "scatteredConfigFlagReads": 0,
